Extract traits from the extension trait
